Verify equal file content before deletion

+++ src/main.rs
@@ -1,7 +1,7 @@
 use std::{
     env,
     fmt::{Debug, Formatter, Result},
-    fs::remove_file,
+    fs::{read, remove_file},
     ops::Add,
 };
 
@@ -87,9 +87,15 @@ fn clear_folder_duplicates(path: &String) -> bool {
         if index < max_index {
             let next_file = dir_files.get(index + 1).unwrap();
             if file.name == next_file.name && file.size == next_file.size {
-                println!("{:?}\n", &next_file);
-                cleared_items += 1;
-                remove_file(&next_file.path).unwrap();
+                
+                let file_content = read(&file.path).unwrap();
+                let next_file_content = read(&next_file.path).unwrap();
+
+                if file_content == next_file_content {
+                    println!("{:?}\n", &next_file);
+                    cleared_items += 1;
+                    remove_file(&next_file.path).unwrap();
+                }
             }
         }
     }
